add fern perms for removing

+++ packages/fern-dashboard/src/app/actions/removeUserFromOrg.ts
@@ -1,11 +1,8 @@
 "use server";
 
+import * as auth0Management from "@/app/services/auth0/management";
+
 import { getCurrentSession } from "../services/auth0/getCurrentSession";
-import {
-  ensureUserBelongsToOrg,
-  getAuth0ManagementClient,
-  invalidateCachesAfterAddingOrRemovingOrgMember,
-} from "../services/auth0/management";
 import { Auth0UserID } from "../services/auth0/types";
 
 export async function removeUserFromOrg({
@@ -13,12 +10,18 @@ export async function removeUserFromOrg({
 }: {
   userIdToRemove: Auth0UserID;
 }) {
-  const auth0 = getAuth0ManagementClient();
+  const auth0 = auth0Management.getAuth0ManagementClient();
   const { userId, orgId } = await getCurrentSession();
-  await ensureUserBelongsToOrg(userId, orgId);
+  await auth0Management.ensureUserBelongsToOrg(userId, orgId);
 
   if (userId === userIdToRemove) {
     throw new Error("User cannot remove themself");
+  }
+
+  const isFernEmployee = await auth0Management.createIsFernEmployee();
+
+  if (!isFernEmployee(userId) && isFernEmployee(userIdToRemove)) {
+    throw new Error("Non-fern-employee cannot remove fern-employee");
   }
 
   await auth0.organizations.deleteMembers(
@@ -26,7 +29,7 @@ export async function removeUserFromOrg({
     { members: [userIdToRemove] }
   );
 
-  await invalidateCachesAfterAddingOrRemovingOrgMember({
+  await auth0Management.invalidateCachesAfterAddingOrRemovingOrgMember({
     userId,
     orgId,
   });
